IGNITE-16691 move page-memory config
