+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/configuration/schema/MemoryAllocatorConfigurationSchema.java
@@ -15,17 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.configuration.schemas.store;
+package org.apache.ignite.internal.pagememory.configuration.schema;
 
-import static org.apache.ignite.configuration.schemas.store.UnsafeMemoryAllocatorConfigurationSchema.UNSAFE_MEMORY_ALLOCATOR_TYPE;
+import static org.apache.ignite.internal.pagememory.configuration.schema.UnsafeMemoryAllocatorConfigurationSchema.UNSAFE_MEMORY_ALLOCATOR_TYPE;
 
-import org.apache.ignite.configuration.annotation.PolymorphicConfigInstance;
+import org.apache.ignite.configuration.annotation.PolymorphicConfig;
+import org.apache.ignite.configuration.annotation.PolymorphicId;
 
 /**
- * Memory allocator that allocates data in offheap using {@link sun.misc.Unsafe}.
+ * Configuration schema for memory allocation strategies.
  */
-@PolymorphicConfigInstance(UNSAFE_MEMORY_ALLOCATOR_TYPE)
-// TODO: IGNITE-16691 Don't forget to redo after IGNITE-16280
-public class UnsafeMemoryAllocatorConfigurationSchema extends MemoryAllocatorConfigurationSchema {
-    public static final String UNSAFE_MEMORY_ALLOCATOR_TYPE = "unsafe";
+@PolymorphicConfig
+public class MemoryAllocatorConfigurationSchema {
+    @PolymorphicId(hasDefault = true)
+    public String type = UNSAFE_MEMORY_ALLOCATOR_TYPE;
 }
